Cahnge epoch_tlim type to float

+++ benchmark_dynamic.py
@@ -26,7 +26,7 @@ def parse_args():
     parser.add_argument(
         "--instance_pattern", default="instances/ORTEC-VRPTW-ASYM-*.txt"
     )
-    parser.add_argument("--epoch_tlim", type=int, default=60)
+    parser.add_argument("--epoch_tlim", type=float, default=60)
 
     return parser.parse_args()
